Module Trace_tef

TEF collector for Trace.

This emits chrome traces (https://docs.google.com/document/d/1CvAClvFfyA5R-PhYUmn5OOQtYMH4h6I0nSsKchNAySU/), which are very simple and a known quantity.

They can be opened in https://ui.perfetto.dev .

module Collector_tef : sig ... end

module Exporter : sig ... end

An exporter, takes JSON objects and writes them somewhere

module Writer : sig ... end

Write JSON events to a buffer.

module Types : sig ... end

type output = [

| `Stdout

| `Stderr

| `File of string

]

Output for tracing.

`Stdout will enable tracing and print events on stdout

`Stderr will enable tracing and print events on stderr

`File "foo" will enable tracing and print events into file named "foo"

val collector : out:[< output ] -> unit -> Trace_core.collector

Make a collector that writes into the given output. See setup for more details.

val setup : ?debug:bool -> ?out:[ output | `Env ] -> unit -> unit

setup () installs the collector depending on out.

parameter out
can take different values:

regular output value to specify where events go

`Env will enable tracing if the environment variable "TRACE" is set.

If it's set to "1", then the file is "trace.json".

If it's set to "stdout", then logging happens on stdout (since 0.2)

If it's set to "stderr", then logging happens on stdout (since 0.2)

Otherwise, if it's set to a non empty string, the value is taken to be the file path into which to write.

parameter debug
if true, use Trace_debug. Default false.

val with_setup : ?debug:bool -> ?out:[ output | `Env ] -> unit -> (unit -> 'a) -> 'a

with_setup () f (optionally) sets a collector up, calls f(), and makes sure to shutdown before exiting. since 0.2 a () argument was added.